requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
openai==1.3.5
PyPDF2==2.11.1
python-docx==0.8.11
nltk==3.6.3
sentence-transformers==2.2.2
huggingface_hub==0.16.4
aiofiles==0.8.0
faiss-cpu==1.7.2
numpy==1.21.0
python-dotenv==0.19.1
torch==1.13.1
transformers==4.30.2